refactor: 상수 추가, 추가된 상수로 기존 교체, 상수 진행하면서 Main 클래스 내 행동의 기존 확률들의 가중치 계산 코드 수정,

+++ src/champion/Sion.java
@@ -15,9 +15,9 @@ public class Sion extends Champion implements Tank, Warrior {
     @Override
     public boolean useQ(Champion target) {
         System.out.printf(getName() + " 이(가) 대량 학살 강타를 사용합니다! 각각 %d%%, %d%%, %d%% 확률 별로 %d, %d, %d의 공격을 가합니다!\n"
-                , (sionQSkillPercent1 * 100)
-                , (sionQSkillPercent2 * 100)
-                , (sionQSkillPercent3 * 100)
+                , (int) (sionQSkillPercent1 * 100)
+                , (int) (sionQSkillPercent2 * 100)
+                , (int) (sionQSkillPercent3 * 100)
                 , (int) Math.round(getAttackPoint() * sionQSkillDamageMultiple1)
                 , (int) Math.round(getAttackPoint() * sionQSkillDamageMultiple2)
                 , (int) Math.round(getAttackPoint() * sionQSkillDamageMultiple3));
